ci: upgrade rust

+++ src/infra/src/file_list/sqlite.rs
@@ -287,11 +287,10 @@ SELECT stream, date, file, deleted, min_ts, max_ts, records, original_size, comp
         time_range: Option<(i64, i64)>,
         flattened: Option<bool>,
     ) -> Result<Vec<FileKey>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
+        if let Some((start, end)) = time_range
+            && start == 0 && end == 0 {
                 return Ok(Vec::new());
             }
-        }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
 
@@ -338,11 +337,10 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         date_range: Option<(String, String)>,
     ) -> Result<Vec<FileKey>> {
-        if let Some((start, end)) = date_range.as_ref() {
-            if start.is_empty() && end.is_empty() {
+        if let Some((start, end)) = date_range.as_ref()
+            && start.is_empty() && end.is_empty() {
                 return Ok(Vec::new());
             }
-        }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
 
@@ -402,11 +400,10 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         time_range: Option<(i64, i64)>,
     ) -> Result<Vec<super::FileId>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
+        if let Some((start, end)) = time_range
+            && start == 0 && end == 0 {
                 return Ok(Vec::new());
             }
-        }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
         let (time_start, time_end) = time_range.unwrap_or((0, 0));
@@ -510,11 +507,10 @@ SELECT id, stream, date, file, deleted, min_ts, max_ts, records, original_size, 
         stream_name: &str,
         time_range: Option<(i64, i64)>,
     ) -> Result<Vec<String>> {
-        if let Some((start, end)) = time_range {
-            if start == 0 && end == 0 {
+        if let Some((start, end)) = time_range
+            && start == 0 && end == 0 {
                 return Ok(Vec::new());
             }
-        }
 
         let stream_key = format!("{org_id}/{stream_type}/{stream_name}");
 
@@ -665,16 +661,16 @@ SELECT stream, MIN(min_ts) as min_ts, MAX(max_ts) as max_ts, COUNT(*) as file_nu
             "#,
         );
         if deleted {
-            sql = format!("{} AND deleted IS TRUE", sql);
+            sql = format!("{sql} AND deleted IS TRUE");
         }
         let sql = match pk_value {
-            None => format!("{} GROUP BY stream", sql),
-            Some((0, 0)) => format!("{} GROUP BY stream", sql),
+            None => format!("{sql} GROUP BY stream"),
+            Some((0, 0)) => format!("{sql} GROUP BY stream"),
             Some((min, max)) => {
                 if deleted {
-                    format!("{} AND id <= {} GROUP BY stream", sql, max)
+                    format!("{sql} AND id <= {max} GROUP BY stream")
                 } else {
-                    format!("{} AND id > {} AND id <= {} GROUP BY stream", sql, min, max)
+                    format!("{sql} AND id > {min} AND id <= {max} GROUP BY stream")
                 }
             }
         };
@@ -702,7 +698,7 @@ SELECT stream, MIN(min_ts) as min_ts, MAX(max_ts) as max_ts, COUNT(*) as file_nu
                 stream_name.unwrap()
             )
         } else {
-            format!("SELECT * FROM stream_stats WHERE org = '{}';", org_id)
+            format!("SELECT * FROM stream_stats WHERE org = '{org_id}';")
         };
         let pool = CLIENT_RO.clone();
         let ret = sqlx::query_as::<_, super::StatsRecord>(&sql)
@@ -721,8 +717,7 @@ SELECT stream, MIN(min_ts) as min_ts, MAX(max_ts) as max_ts, COUNT(*) as file_nu
         stream_name: &str,
     ) -> Result<()> {
         let sql = format!(
-            "DELETE FROM stream_stats WHERE stream = '{}/{}/{}';",
-            org_id, stream_type, stream_name
+            "DELETE FROM stream_stats WHERE stream = '{org_id}/{stream_type}/{stream_name}';"
         );
         let client = CLIENT_RW.clone();
         let client = client.lock().await;
@@ -955,8 +950,8 @@ UPDATE stream_stats
         };
         let id = ret.try_get::<i64, &str>("id").unwrap_or_default();
         let status = ret.try_get::<i64, &str>("status").unwrap_or_default();
-        if id > 0 && super::FileListJobStatus::from(status) == super::FileListJobStatus::Done {
-            if let Err(e) =
+        if id > 0 && super::FileListJobStatus::from(status) == super::FileListJobStatus::Done
+            && let Err(e) =
                 sqlx::query("UPDATE file_list_jobs SET status = $1 WHERE status = $2 AND id = $3;")
                     .bind(super::FileListJobStatus::Pending)
                     .bind(super::FileListJobStatus::Done)
@@ -969,7 +964,6 @@ UPDATE stream_stats
                 }
                 return Err(e.into());
             }
-        }
         if let Err(e) = tx.commit().await {
             log::error!("[SQLITE] commit add job error: {e}");
             return Err(e.into());
